Obstacle avoidance
Completed challenge 2. It works most of the time, but sometimes it doesn't. Sometimes it gets stuck in a loop trying to find the nearest vertex. Sometimes in plotting the path to the goal it skips several vertices. Updated the empty_world.py script so that dq will affect the exploration rate.